Add custom token pool utilities (#342)
* test runner: add SyncTokenPools command to enable add/removePool

A new `syncTokenPools` is added to the `TestCCIP` script, which allows
syncing the registered tokenPools from EVMChainConfig to the on-chain
smartcontracts. So, in order to add a new pool to the onRamp/offRamp on
a given network, simply add the token and pool addresses to the rhea
EVMChainConfig, and call TestCCIP with COMMAND=syncTokenPools.

* ccip: add BurnMintTokenPool contract

* ccip: make SyncTokenPools also setPrices, deduplicate Source/Dest code

A new EvmChainConfig.TokenPrices config parameter is added, which
should be either an empty slice, or uint256s corresponding to the prices
to be set for each corresponding BridgeTokens. Also, we define the
tokenPoolRegistry and aggregateRateLimiter interfaces, which are
implemented by OnRamp/OffRamp, and use them to deduplicate code for
syncPool and syncPrices.

* ccip: BurnMintTokenPool test with one of existing token pools

* syncTokenPools: hardcode gasLimit and set offramp to source tokens

* add snxUSD BurnMintTokenPools addresses to Prod and Staging_Beta envs

* BurnMintTokenPool: add gethwrappers

* chore: generate burn mint wrapper

Co-authored-by: Rens Rooimans <github@rensrooimans.nl>