Moving DynamicConfig to the ocr2.onChainConfig (#689)
Instead of building a custom solution on top of event polling, we want to move setting DynamicConfig from both contracts to onChainConfig from the OCR2 base contract. This approach will give us an automatic plugin restart when the configuration changes. By default, onChainConfig is stored as a byte slice within OCR2Base.sol. Additionally, it will be kept as a private field to make reading it from contracts efficient.

Missing remarks will be applied in a separate PR