fix: incorrect transferFrom logic
The Rebalancer contract will always call the bridge adapter like so:

```solidity
remoteLiqManager.localBridge.sendERC20{value: nativeBridgeFee}(
    address(i_localToken),
    remoteLiqManager.remoteToken,
    remoteLiqManager.remoteRebalancer,
);
```

however the IBridgeAdapter implementations sometimes do a
transferFrom from the 2nd argument and sometimes from the 1st argument,
which is incorrect behavior.

To remedy this, name IBridgeAdapter arguments to be local/remote instead,
which matches the nomenclature of the Rebalancer contract itself, and switch
all transferFrom's to transferFrom the local token, which is the token where
all the approvals are made to.