Make deviation PPB calc based on smaller value (#482)
We typically check deviation by `deviates(newPrice, lastPrice)`,
`deviates` function used the 1st param, i.e. newPrice, as the
denominator, making deviation >100% impossible if price is increasing.
E.g., newPrice = 1000, lastPrice is 1, then deviation is calculated as
(1000-1)/1000 = 99.9%. This means if we have deviations threshold set to
> 100%, like we have on mainnet, then it will never deviate if price is
increasing.

make deviation calculation always be based on the smaller value, so that
deviation result is consistent whether price is increasing or
decreasing.

If price is increasing, fixes the issue when threshold is >100%, and
makes price update more frequent
If price is decreasing, no behavior change