Exec NewReportingPlugin must not error (#894)
Previously, any single error during the call to NewReportingPlugin for
the Exec plugin would cause the OCR instance to not start. Even if the
error was transient and would be resolved if NewReportingPlugin was
called again, the OCR instance would remain down.

In order to be robust to transient errors, NewReportingPlugin is retried
until it succeeds.