core/capabilities/ccip: fix oracle startup logic
Ticket: https://smartcontract-it.atlassian.net/browse/CCIP-3078

Remove the bootstrap oracle startup logic. This is
actually not needed since we specify V2Bootstrappers
on the libocr3.OracleArgs struct, since this will
cause the created oracle to automatically bootstrap
without the need for the bootstrap node.

This does mean that it is vital that the P2PV2Bootstrappers
field in the CCIP capability job spec is correctly set.

This also speeds up tests mildly by about 20% which
is a good quality of life bonus.