Zp/update mockrouter (#1427)
Specifically, allow for empty message with zero gas.

## Motivation
EVM2EVMOffRamp checks the message for content and gaslimit and skips
calling ccip receive if they're not present.
Consequently the mock errors (ReceiverError) when sending only tokens to
a smart contract. This update will help chainlink-local correctly
process fork-based tests too.

## Solution
The MockRouter implementation of _routeMessage was last updated in
https://github.com/smartcontractkit/ccip/pull/669 and is out of sync
with the logic in EVM2EVMOffRamp.