CRIB-359: ignoring problematic git tag breaking CRIB CCIP builds
Fix the following error when building CCIP for CRIB:

```
build:app   • parsing tag
build:app   ⨯ release failed after 0s                          error=failed to parse tag 'contracts-ccip/v1.5.0-beta.0' as semver: Invalid Semantic Version
```

Even though the best would be to ignore it based on the prefix `contracts-ccip`, that requires goreleaser-pro
which is not currently available in the nix shell. We should add it in the long run, but this issue needs an immediate fix for now.