BCFR-1076 false unreachable RPC transition on rotation (#1548)
## Motivation
In cases when lease duration > 0 MultiNode periodically rotates RPCs. As
we have not marked the finalized head subscription as part of
aliveLoopSub on RPC rotation, the subscription was closed, which caused
RPC to transition to an unreachable state.

## Solution

Mark finalized subscription as part of alive loop to avoid subscription
termination on RPC rotation.